[test] Move some tests from qt -> wallet
After previous refactoring, the tests make more sense here.
